Move api package to lib
