+++ pkg/lib/sms/accessyou/smsclient.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"regexp"
 
-	"github.com/authgear/authgear-sms-gateway/pkg/lib/sms/api"
+	"github.com/authgear/authgear-sms-gateway/pkg/lib/api"
 	"github.com/authgear/authgear-sms-gateway/pkg/lib/sms/smsclient"
 )
