+++ pkg/lib/sms/smsclient/smsclient.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/authgear/authgear-sms-gateway/pkg/lib/api"
 	"github.com/authgear/authgear-sms-gateway/pkg/lib/sensitive"
-	"github.com/authgear/authgear-sms-gateway/pkg/lib/sms/api"
 )
 
 type TemplateVariables struct {
